Button press recognition code samples
The polling approach to button press recognition (on the ATmega88) seems to be working very well, but debugging still needs to be done on the external interrupt approach code. In that code, nothing is happening when the button is pressed; I am receiving a warning from avr-gcc that "‘PCINT22_vect’ appears to be a misspelled signal handler [enabled by default]" when I compile my code, so I will do some further investigation on that. JGarcia