The Project Directories for the ATmega88/328/32
I'm continuing work on the modification of Stefan's NRF24L01 library (mostly working on the copy in the INTERCOM32 directory). In that directory, I have also included a more fleshed out addressing plan, as well as listing which microcontroller will go in which location in the house. I am hoping to get a basic link going soon!